backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]
bcrypt==4.3.0
pydantic-settings==2.1.0
fastapi-mail==1.4.1
python-multipart==0.0.6
email-validator==2.1.0
slowapi==0.1.9
psycopg2-binary>=2.9.0